zrpc: remove blocking call when reading thrift
Following side effect can be observed. When both updater and
configurator threads are running, if configurator is pending, then
it must finish so as to permit updater to be launched. The fix
consists in removing call to transport_peek() that is a function
that blocks other connections and it is useless, because of thread
mechanism implemented.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>