zrpcd: sanity check when reading capnproto context for getting routes
Because the same target structure is used to get information from
capnproto, if it is called twice, then some contexts may be lost,
like capnproto pointers or text structures that require specific
allocation. This fix consists in avoiding twice the same call, since
it is useless, and it avoids overwriting an already allocated pointer.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>