zrpc: duplicate and move qzc.capnp and qzmq file from lib
A part of the functions defined in those 4 files are used as part of the
communication protocol between zrpcd and bgpd. The choice to
copy the files has been done in order to better identify the functions
to be used. The choice has been done also, in order to be sure that
zrpcd does not use other files than the ones contained in that
folder.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>