zrpc: enhance utilities function for zrpc daemon
This set of internal APIs are used for manipulation of objects received
and to be sent to capnproto interface. Among functions useful, one can
find how to translate route targets into uint64_t lists, how to fill in
internal structure for route target, and how to compare route targets.
There is also an API to get the path where bgpd daemon is. Actually,
bgpd is launched by zrpc daemon.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>