zrpc: introduction of bgp capnproto code inherited from bgp.capnp file
This commit introduces the bgp.capnp file that can be found also on
quagga BGP folder. In addition to this, other changes include files
that are built based on structure definition of that bgp.canp file
As the c_generation based on the scheme is not available, the header and
source file have been done. One can also find some defines useful for
filling in the various bgp structures exchanged between the two
applications.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>