zrpcd: enhance esi and mac handling functions
Enhance zrpc routines used to manipulate ESI structure and MAC
structure. Those functions are used to check for input ESI and mac
address format. Only accepted format is <aa:bb:cc:..:ff> where the
number of elements is 6 for mac address, and 10 for ESI. Those functions
are also used to internally store the ESI and mac address structure.
There is also a function to extract and append a mac address to an
extended community attribute.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>