zrpcd: monitor launched BGPd, in case of failure
In case bgp daemon disappears abruptly, it has not the time to clean all
the context, including the capnproto context exchange. This is up to the
zrpc daemon to detect that the connection has been broken. This is done,
by getting the SIGCHLD signal, and cleaning the context in relation to
the bgp daemon.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>