zrpcd: support L3VPN dual stack
To support dual stack, two parameters (afi and safi) are added for the two thrift API:
addVrf and delVrf. This commit handles the two paramters for addVrf and delVrf.

To support dual stack, use parameter --with-thrift-version=5 when running configure
programs. The dual stack thrift model is based on thrift model version 4.

Signed-off-by: Liu Xiaofeng <xiaofeng.liu@6wind.com>