zrpcd: detect if thrift message has been sent
This change detects if an error from thrift layer has been detected when
sending thrift message. This occurs when an error stream is generated,
and is related to SEND ERROR message. With that, the number of lost
messages is accounted, and an error message is displayed on the log
file. Also, the retry timer is immediately restarted.
Also, to bring less confusion, the error message signaling the change of
state on the 6644 connection is moved from debug trace to error trace.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>