zrpcd: change thrift bgp updater call to non blocking
Enhancement so that the connection from QBGP to ODL on port 6644 will be
retried each second if there is a failure in receiving any responses
from remote.
Previously, the thrift_transport_open() call was blocking, when the
connect() call, and this call could block if no response from the remote
was happening. Now a replacement of thrift_transport_open() is done, and
this call will configure 6644 client connection to non blocking mode.
If the primary call() fails because of EINPROGRESS errno, then one second later,
a select() call will be done to check if the connection attemps was
success or not. If not, the socket will be reset, and a new attempt will
be triggered. If yes, then subsequent exchanges between ODL and QBGP
will be done.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>