zrpcd: allow configurator server listening on specified address
In commit "fdab7218 zrpcd: add zrpc config options to restrict listen
port and address", specified address and port are required to be
bound to the listening socket. But the commit doesn't get to the
purpose, because there is no 'hostname' member in struct
'ThriftServerSocket':

   setup->bgp_configurator_server_transport =
     g_object_new (THRIFT_TYPE_SERVER_SOCKET,
                        "hostname",  tm->zrpc_listen_address,
                        "port", setup->zrpc_listen_port,
			                   NULL);
So the above codes does no help.

The function thrift_server_socket_listen() is called by zrpcd to
create the listening socket. There is a disavantage that zrpcd cannot
listen on specified address. Now zrpc_server_socket_listen() is defined
to re-implement thrift_server_socket_listen(). In the new function, the
address specified by "-L" parameter will be bound to the listening socket.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>