zrpcd: introduce a retry algorithm to send thrift packets to remote
There are cases where thrift can not assume sending thrift information
to remote, because of a transportation failure. The failure may be due
to socket busy, which means that the socket needs to be dequeued to be
used.

This commit introduces a retry algorithm to send thrift packets to
remote in case of failure. Also statistics on retries thrift emissions
are counted and displayed.

zrpcd can be called with -s option, by passing a value that is the
max timeout value to be used. The default value for select timeout is
10 seconds.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>