zrpcd: fix unexpected error in getRoutes handler
If there are more than 2 vrfs, a following error message will
be printed in logs when getRoutes() handler is called:

  Debug ZRPC: qzcclient_send. reply message error: (1)

A loop on vrfs is to parse all routes per vrf. After the 1st vrf
is parsed in the first loop, the 2nd vrf should be parsed in the
second loop. But now still the 1st vrf is parsed while it will
be removed from ctxt->bgp_get_routes_list. So error occurs.

This commit fixes this issue. Also, useless lines are deleted.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>