zrpcd: add a framework to support stale config
This commit introduces a framework to support stale config.

In case of breaking of the connection from qbgp to odl, this may
signify that an odl move within a cluster may take place; a timer
is run, in order to remember to remove the non unmarked entries
(vrfs, peers, and prefixes). That timer will generally stick with
end of rib, that is to say either the one defined by sendEorDelay()
or when sendEor() is called. In order to give more flexibility,
the timer by default is set to 1800 Seconds maximum (which is
supposed to be the same as Eor delay), but can also be changed,
by using option '-M <1-3600>' value.

If the stalemarker timer is set to 0, it is meaning that the
configuration is kept indefinitely.

The stalemarking is launched when TCP connection on port 7644
from ODL is re-established. When one vrf/peer/prefix is added
again, the stalemarking is unset.

Upon sendEor() call, all vrfs/peers/prefixes marked by stale flag
are flushed.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>