zrpcd: stalemarker config support for VRF and BGP peer
This commit adds support for marking all bgp vrfs and peers, once
connection to sdn controller is broken.

At that time, a stale timer is run. The staled vrfs and peers will
be flushed when sendEor() is received. Otherwise an error message
is logged upon the stale timer expiration, without flushing the
staled vrfs and peers.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>