zrpcd: set ZMQ_LINGER to 0 for new created zmq socket
The default socket's linger period is "infinite", by default the
function zmq_ctx_term() will block indefinitely if there are any
pending connects or sends. Accoding to the doc of zmq, it is
strongly recommended to set ZMQ_LINGER to zero on all sockets
and close all sockets, before calling zmq_ctx_term().

Even with that, zmq_ctx_term() would still block in the case that
a zmq user was sending messages to a peer that was not running at
all. This commit sets ZMQ_LINGER option with 0 for new created zmq
socket to avoid that case.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>